+++ impactlab_tools/__init__.py
@@ -5,7 +5,7 @@
 from __future__ import absolute_import
 
 __author__ = """Climate Impact Lab"""
-__version__ = '0.2.0'
+__version__ = '0.3.1'
 
 _module_imports = (
 )